refactor: 스토리지 업로드, svg 컨버트를 위한 빈 클래스를 인터페스화
