+++ backend/src/main/java/com/woowacourse/zzimkkong/infrastructure/StorageUploader.java
@@ -1,0 +1,7 @@
+package com.woowacourse.zzimkkong.infrastructure;
+
+import java.io.File;
+
+public interface StorageUploader {
+    String upload(final String directoryName, final File uploadFile);
+}
