+++ backend/src/main/java/com/woowacourse/zzimkkong/service/MapService.java
@@ -10,7 +10,7 @@ import com.woowacourse.zzimkkong.dto.map.MapFindResponse;
 import com.woowacourse.zzimkkong.exception.authorization.NoAuthorityOnMapException;
 import com.woowacourse.zzimkkong.exception.map.NoSuchMapException;
 import com.woowacourse.zzimkkong.exception.space.ReservationExistOnSpaceException;
-import com.woowacourse.zzimkkong.infrastructure.S3Uploader;
+import com.woowacourse.zzimkkong.infrastructure.StorageUploader;
 import com.woowacourse.zzimkkong.infrastructure.SvgConverter;
 import com.woowacourse.zzimkkong.infrastructure.TimeConverter;
 import com.woowacourse.zzimkkong.repository.MapRepository;
@@ -28,7 +28,7 @@ public class MapService {
     private final MapRepository maps;
     private final SpaceRepository spaces;
     private final ReservationRepository reservations;
-    private final S3Uploader s3Uploader;
+    private final StorageUploader storageUploader;
     private final SvgConverter svgConverter;
     private final TimeConverter timeConverter;
 
@@ -36,13 +36,13 @@ public class MapService {
             final MapRepository maps,
             final SpaceRepository spaces,
             final ReservationRepository reservations,
-            final S3Uploader s3Uploader,
+            final StorageUploader storageUploader,
             final SvgConverter svgConverter,
             final TimeConverter timeConverter) {
         this.maps = maps;
         this.spaces = spaces;
         this.reservations = reservations;
-        this.s3Uploader = s3Uploader;
+        this.storageUploader = storageUploader;
         this.svgConverter = svgConverter;
         this.timeConverter = timeConverter;
     }
@@ -119,7 +119,7 @@ public class MapService {
 
     private String uploadPngToS3(final String svgData, final String fileName) {
         File pngFile = svgConverter.convertSvgToPngFile(svgData, fileName);
-        String thumbnailUrl = s3Uploader.upload("thumbnails", pngFile);
+        String thumbnailUrl = storageUploader.upload("thumbnails", pngFile);
         pngFile.delete();
         return thumbnailUrl;
     }
