+++ backend/src/test/java/com/woowacourse/zzimkkong/service/ServiceTest.java
@@ -1,6 +1,7 @@
 package com.woowacourse.zzimkkong.service;
 
 import com.woowacourse.zzimkkong.infrastructure.S3Uploader;
+import com.woowacourse.zzimkkong.infrastructure.StorageUploader;
 import com.woowacourse.zzimkkong.repository.MapRepository;
 import com.woowacourse.zzimkkong.repository.MemberRepository;
 import com.woowacourse.zzimkkong.repository.ReservationRepository;
@@ -27,5 +28,5 @@ public class ServiceTest {
     protected SpaceRepository spaces;
 
     @MockBean
-    protected S3Uploader s3Uploader;
+    protected StorageUploader storageUploader;
 }
